// frontend/src/model/User.model.ts
export interface User{
    id: number;
    image:string;
    userName: string;
    firstName:string;
    lastName:string;
    phoneNumber:string;
    place:string;
    email: string;
    emailSubscription: string;
    password: string;
    confirmPassword: string;
    registrationDate: string; 
    university:string;
    work:string;


}